move lots of general settings to the job options structure that can be set at the global, queue, or job level all settings work except for prunedone TODO: make prunedone work TODO: add missing settings for cron jobs TODO: document new settings

+++ src/brooce/web/tpl/tpl.go
@@ -49,7 +49,7 @@ func Get() *template.Template {
 
 			return util.HumanDuration(time.Unix(end, 0).Sub(time.Unix(start, 0)), 1)
 		},
-		"TimeDuration": func(seconds int64) string {
+		"TimeDuration": func(seconds int) string {
 			if seconds == 0 {
 				return ""
 			}
